Split host message switch into handler registry (oh-tab-zhwi.4.2)
